add profile based claims validation
Also split Claims and Evidence into different files
